requirements.txt
brill_postagger
unicode_rbnf